refactor: add missing imports (part 6)
