+++ src/renderer/app/components/Titlebar/index.tsx
@@ -5,14 +5,14 @@ import { ipcRenderer } from 'electron';
 import { WindowsControls } from 'react-windows-controls';
 
 import store from '~/renderer/app/store';
+import { ToolbarButton } from '~/renderer/components/ToolbarButton';
+import { StyledTitlebar, TrafficButtons, Handle, Icon, Title } from './style';
+import { icons } from '~/renderer/constants/icons';
 import {
   closeWindow,
   maximizeWindow,
   minimizeWindow,
-} from '~/renderer/app/utils';
-import { ToolbarButton } from '~/renderer/components/ToolbarButton';
-import { icons } from '~/renderer/constants';
-import { StyledTitlebar, TrafficButtons, Handle, Icon, Title } from './style';
+} from '../../utils/window';
 
 const onUpdateClick = () => {
   ipcRenderer.send('update-install');
